add create word, upd name WordController->HomeController, add new WordController
