+++ resources/views/site/create.blade.php
@@ -1,0 +1,22 @@
+@extends('layouts.app')
+
+@section('content')
+
+
+<div class="bg-white p-4">
+    <div class="d-flex justify-content-between align-items-center mb-2">
+        <h4>Create word</h4>
+    </div>
+
+    <form action="{{ route('word.store') }}" method="post">
+        @csrf
+        <label for="word">Word:</label>
+        <input type="text" name="word" required>
+        <button class="btn btn-outline-dark" type="submit">Create</button>
+
+    </form>
+
+</div>
+
+
+@endsection
